pyproject.toml
[project]
name = "serpents-per-second"
version = "1.0.0"
description = "SNES-styled Snake game with multiple modes, high scores, and chiptune audio"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Author", email = "author@example.com"}
]
dependencies = [
    "pygame-ce>=2.5.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.9",
    "mypy>=1.11.2",
    "pillow>=10.4.0",
    "pyinstaller>=6.3.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "src/app.py",
    "src/scenes/*",
    "src/render/*",
    "src/systems/audio.py",
    "src/systems/scaling.py",
    "src/systems/logging_setup.py",
]

[tool.coverage.report]
fail_under = 70
show_missing = true

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
